Windows/MSVC: enable curated core test suite (type/vector/base) + portability fixes
Enable VELOX_BUILD_TESTING on Windows and get a curated set of core test
executables building and passing under MSVC: velox_type_test (273 pass),
velox_vector_test (479 pass) and velox_base_test (160 pass) -- 912 cases green.

Build-system enablement:
- duckdb.cmake / apply_patches.cmake: apply DuckDB patches via cmake -P (no
  git apply on Windows); static-lib DUCKDB_API, drop ccache, re2 PRIVATE include.
- gtest.cmake: Windows build wiring for the test framework.
- CMakeLists.txt: suppress MSVC C4716 (noreturn throw helpers) for /W3 test targets.

Core portability bugs found via the tests (LLP64: long/UL are 32-bit on Windows):
- Nulls.h: kNotNull64 = ~0UL -> ~0ULL (was 0xFFFFFFFF, not all-ones 64-bit).
- type/windows/Int128.h: Int128 toAppend now emits decimal (matching POSIX
  folly::to<__int128>) instead of hex.
- Type.cpp: IntervalDayTimeType::valueToString casts the int128 remainder to
  int64_t before snprintf (passing the non-trivial Int128 class to varargs is UB).
- Subfield.h / Tokenizer.cpp: LongSubscript index and parsing use int64_t /
  std::stoll instead of long / std::stol (std::stol overflows 32-bit on Windows).
- FlatVector.cpp: restore FlatVector<bool>::set specialization (present in
  upstream base, dropped by an earlier port commit; needed by vector tests).

Test-only MSVC portability fixes (POSIX paths kept byte-identical via guards):
- TimestampTest.cpp: full-range gmtime_r/timegm shims (Howard Hinnant civil-day
  algorithms mirroring epochToCalendarUtc) and a putTimePortable helper, since
  MSVC _gmtime64_s/_mkgmtime and std::put_time choke on extreme/negative years.
- BitUtilTest.cpp: __builtin_popcountl -> ...ll and __uint128_t -> uint128_t alias.
- ConcurrentCounterTest / AssertQueryBuilder / VectorTestBase: folly
  hardware_concurrency() fallback (vcpkg folly predates available_concurrency()).
- SplitBlockBloomFilterTest.cpp: bzero -> std::memset.
- DuckConversion.h, BiasVector-inl.h, VectorFuzzer.cpp, TimeZoneDatabase.cpp and
  assorted type/vector test files: explicit int128 narrowing, if-constexpr guards,
  and other minor _WIN32/_MSC_VER-guarded adjustments.

+++ velox/type/Type.cpp
@@ -1377,7 +1377,7 @@ std::string IntervalDayTimeType::valueToString(int64_t value) const {
       hours,
       minutes,
       seconds,
-      remainMillis);
+      static_cast<int64_t>(remainMillis));
 
   return buf;
 }
